Created custom logicalTypes and conversions JSONDATE JSONDECIMAL JSONTIME_MICROS JSONLOCAL_TIMESTAMP_MICROS JSONLOCAL_TIMESTAMP_MILLIS JSONTIME_MILLIS JSONTIMESTAMP_MICROS JSONTIMESTAMP_MILLIS
This to assure the genericdatawriter and jsonencoder that are used convert these logicaltypes to string values instead of long, int and bytes as the genericdatumwriter instructs to do per default.

createLogicalTypesStringSchema method that creates a target schema from the original schema that replaces the equivalent avro logicaltypes with the respective JSON* logical datatypes. This target schema is then used by GenericDatumWriter when serializing the payload to the payload.field.name as JSON string.